Sorted through links
A dilema early on: this algorithm isn't giving the full results. As this stands, this would be beneficial to many teams eyeing a chance to qualify for the state meet (using results from athletic.net), but not so much for teams towards the bottom of the list. I'll consider using the Hypothetical Meet option later, to see if that yields any different/better results overall.